+++ brainpy/backend/__init__.py
@@ -3,8 +3,8 @@
 from types import ModuleType
 
 from brainpy import errors
-from .operators.bk_numpy import *
-from .runners.general_runner import GeneralNodeRunner, GeneralNetRunner
+from .ops.numpy_ import *
+from .runners.general import GeneralNodeRunner, GeneralNetRunner
 
 
 BUFFER = {}
@@ -57,57 +57,57 @@ def set(backend=None, module_or_operations=None, node_runner=None, net_runner=No
 
     global_vars = globals()
     if backend == 'numpy':
-        from .operators import bk_numpy
+        from .ops import numpy_
 
         node_runner = GeneralNodeRunner if node_runner is None else node_runner
         net_runner = GeneralNetRunner if net_runner is None else net_runner
-        module_or_operations = bk_numpy if module_or_operations is None else module_or_operations
+        module_or_operations = numpy_ if module_or_operations is None else module_or_operations
 
     elif backend == 'pytorch':
-        from .operators import bk_pytorch
+        from .ops import pytorch_
 
         node_runner = GeneralNodeRunner if node_runner is None else node_runner
         net_runner = GeneralNetRunner if net_runner is None else net_runner
-        module_or_operations = bk_pytorch if module_or_operations is None else module_or_operations
+        module_or_operations = pytorch_ if module_or_operations is None else module_or_operations
 
     elif backend == 'tensorflow':
-        from .operators import bk_tensorflow
+        from .ops import tensorflow_
 
         node_runner = GeneralNodeRunner if node_runner is None else node_runner
         net_runner = GeneralNetRunner if net_runner is None else net_runner
-        module_or_operations = bk_tensorflow if module_or_operations is None else module_or_operations
+        module_or_operations = tensorflow_ if module_or_operations is None else module_or_operations
 
     elif backend == 'numba':
-        from .operators import bk_numba_cpu
-        from .runners.numba_cpu_runner import NumbaCPUNodeRunner, set_numba_profile
+        from .ops import numba_cpu
+        from .runners.numba_cpu import NumbaCPUNodeRunner, set_numba_profile
 
         node_runner = NumbaCPUNodeRunner if node_runner is None else node_runner
-        module_or_operations = bk_numba_cpu if module_or_operations is None else module_or_operations
+        module_or_operations = numba_cpu if module_or_operations is None else module_or_operations
         set_numba_profile(parallel=False)
 
     elif backend == 'numba-parallel':
-        from .operators import bk_numba_cpu
-        from .runners.numba_cpu_runner import NumbaCPUNodeRunner, set_numba_profile
+        from .ops import numba_cpu
+        from .runners.numba_cpu import NumbaCPUNodeRunner, set_numba_profile
 
         node_runner = NumbaCPUNodeRunner if node_runner is None else node_runner
-        module_or_operations = bk_numba_cpu if module_or_operations is None else module_or_operations
+        module_or_operations = numba_cpu if module_or_operations is None else module_or_operations
         set_numba_profile(parallel=True)
 
     elif backend == 'numba-cuda':
         raise NotImplementedError
-        from .operators import bk_numba_cuda
-        from .runners.numba_cuda_runner import NumbaCudaNodeRunner
+        from .ops import numba_cuda
+        from .runners.numba_cuda import NumbaCudaNodeRunner
 
         node_runner = NumbaCudaNodeRunner if node_runner is None else node_runner
-        module_or_operations = bk_numba_cuda if module_or_operations is None else module_or_operations
+        module_or_operations = numba_cuda if module_or_operations is None else module_or_operations
 
     elif backend == 'jax':
         raise NotImplementedError
-        from .operators import bk_jax
-        from .runners.jax_runner import JaxRunner
+        from .ops import jax_
+        from .runners.jax_ import JaxRunner
 
         node_runner = JaxRunner if node_runner is None else node_runner
-        module_or_operations = bk_jax if module_or_operations is None else module_or_operations
+        module_or_operations = jax_ if module_or_operations is None else module_or_operations
 
     else:
         if module_or_operations is None:
@@ -232,7 +232,7 @@ def get_node_runner():
     """
     global _node_runner
     if _node_runner is None:
-        from .runners.general_runner import GeneralNodeRunner
+        from .runners.general import GeneralNodeRunner
         _node_runner = GeneralNodeRunner
     return _node_runner
 
@@ -247,7 +247,7 @@ def get_net_runner():
     """
     global _net_runner
     if _net_runner is None:
-        from .runners.general_runner import GeneralNetRunner
+        from .runners.general import GeneralNetRunner
         _net_runner = GeneralNetRunner
     return _net_runner
