+++ brainpy/backend/ops/numba_overload.py
@@ -1,12 +1,15 @@
 # -*- coding: utf-8 -*-
 
+import numpy
+
 from brainpy import errors
 
 try:
     import numba
 except ModuleNotFoundError:
     raise errors.PackageMissingError(errors.backend_missing_msg.format(bk='numba'))
-import numpy
+
+
 from numba.extending import overload
