+++ brainpy/tools/__init__.py
@@ -3,3 +3,4 @@
 from .ast2code import *
 from .codes import *
 from .dicts import *
+from .decorators import *
